🚚 Renamed `Footer` layout component
